requirements.txt
# ==========================================
# AI-MedicalCenter-LangGraph Dependencies
# ==========================================

# Core Flask Framework
flask[async]>=3.0.0,<4.0.0
flask-session>=0.6.0
python-dotenv>=1.0.1
gunicorn>=21.2.0

# LangChain Ecosystem (pinned for stability)
langchain>=0.3.0,<0.4.0
langchain-core>=0.3.0,<0.4.0
langchain-community>=0.3.0,<0.4.0
langchain-openai>=0.3.0,<0.4.0
langchain-groq>=0.3.0,<0.4.0

# LangGraph (Latest compatible versions)
langgraph>=0.4.0,<0.5.0
langgraph-checkpoint>=2.0.0,<3.0.0
langgraph-sdk>=0.1.0,<0.2.0
langgraph-prebuilt>=0.1.0,<0.2.0
langsmith>=0.1.0,<0.2.0

# LLM Providers
openai>=1.50.0,<2.0.0
groq>=0.4.0,<1.0.0

# PDF Generation (comprehensive solution)
reportlab>=4.0.0,<5.0.0
weasyprint>=60.0.0,<61.0.0
pydyf>=0.10.0,<0.11.0
fpdf2>=2.8.0,<3.0.0
pdfkit>=1.0.0,<2.0.0

# Core Utilities
pydantic>=2.10.0,<3.0.0
tenacity>=8.5.0
aiohttp>=3.10.0,<4.0.0
requests>=2.32.0,<3.0.0
tqdm>=4.60.0

# Data Processing
numpy>=1.24.0,<2.0.0
tiktoken>=0.8.0
orjson>=3.10.0
PyYAML>=6.0.2
Pillow>=10.4.0

# Web Framework Utilities
jinja2>=3.1.0,<4.0.0
markupsafe>=2.1.0,<3.0.0
werkzeug>=3.0.0,<4.0.0

# Development and Testing (optional)
pytest>=8.0.0;python_version>="3.8"
pytest-asyncio>=0.23.0;python_version>="3.8"
black>=24.0.0;python_version>="3.8"

# Security and Performance
cryptography>=42.0.0
colorama>=0.4.6

# Additional utilities for medical system
python-dateutil>=2.8.0
uuid>=1.30

# ==========================================
# Advanced Psychology Module Dependencies
# ==========================================

# Statistical Analysis and Data Science
scipy>=1.11.0,<2.0.0
scikit-learn>=1.5.0,<2.0.0
pandas>=2.2.0,<3.0.0
statsmodels>=0.14.0,<0.15.0

# Natural Language Processing for Emotion Analysis
nltk>=3.8.0,<4.0.0
textblob>=0.18.0,<0.19.0
spacy>=3.7.0,<4.0.0
vaderSentiment>=3.3.0,<4.0.0

# Machine Learning for Personality Analysis
joblib>=1.4.0,<2.0.0
plotly>=5.22.0,<6.0.0
matplotlib>=3.9.0,<4.0.0
seaborn>=0.13.0,<0.14.0

# Time Series Analysis for Longitudinal Tracking
matplotlib>=3.9.0,<4.0.0
plotly>=5.22.0,<6.0.0

# Data Validation and Serialization
marshmallow>=3.23.0,<4.0.0
dataclasses-json>=0.6.0,<0.7.0

# Psychological Assessment Tools (implemented directly in code) 